drag and drop fix

+++ src/apollo/mutations.tsx
@@ -2,7 +2,7 @@ export const createProject = `mutation CreatProject($name:String!){
   createProject(name:$name)
 }`;
 
-export const updateIssue = `mutation UpdateIssue($id: ID!, $status: IssueStatus!, $boardPosition: Int!) {
+export const updateIssue = `mutation UpdateIssue($id: ID!, $status: IssueStatus!, $boardPosition: Float!) {
   updateIssue(id: $id, status: $status, boardPosition: $boardPosition)
 }`;
